git-grope: operands name files to grope
$(git ls-files) is still the default

this turned up a bug in git-grope (and 5050.t) which caused it
to ignore its operands (beyond the first one which is treated as
a pattern in absence of `-e PATTERN`.  5050.t asserted results
of the effective `git-grope natural` instead of the intended
`git-grope 'natural|language'`.

would have done this as `$gropetool "${(@)${(A):--e}:^^terms}"`,
unfortunately ${name:^^arrayname} doesn't work in SLE-12's zsh.